Move shared helper tests from plugin suite to shared tests
