+++ package.json
@@ -36,7 +36,8 @@
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "tree": "find . | grep -v \"__pycache__\\|\\.pytest_cache\\|\\.benchmarks\\|node_modules\\|\\.pyc\\|\\.antlr\\|./talk-venv\\|./.github\\|./.DS_Store\\|package-lock.json\\|./.git\" > project_structure.txt",
-    "test": "npm run test:parser && npm run test:plugin",
+    "test": "npm run test:shared && npm run test:parser && npm run test:plugin",
+    "test:shared": "node --test src/shared/tests/*.test.js",
     "test:parser": "cd src/parser && npm test",
     "format:gml": "node ./src/plugin/prettier-wrapper.js",
     "format": "npm exec -- prettier --ignore-path ./.prettierignore --write \"{src/**/*.js,src/**/*.mjs,scripts/**/*.{js,mjs},*.{js,cjs,mjs}}\"",
